removing proxy classes
