+++ sae_lens/saes/standard_sae.py
@@ -1,13 +1,13 @@
-from typing import Any, Optional, Tuple
-
+import numpy as np
 import torch
 from jaxtyping import Float
+from numpy.typing import NDArray
 from torch import nn
 
-from sae_lens.saes.sae_base import BaseSAE, BaseTrainingSAE, SAEConfig
+from sae_lens.saes.sae import SAE, SAEConfig, TrainingSAE, TrainStepInput
 
 
-class StandardSAE(BaseSAE):
+class StandardSAE(SAE):
     """
     StandardSAE is an inference-only implementation of a Sparse Autoencoder (SAE)
     using a simple linear encoder and decoder.
@@ -20,6 +20,8 @@ class StandardSAE(BaseSAE):
     The BaseSAE.forward() method automatically calls encode and decode,
     including any error-term processing if configured.
     """
+
+    b_enc: nn.Parameter
 
     def __init__(self, cfg: SAEConfig, use_error_term: bool = False):
         super().__init__(cfg, use_error_term)
@@ -80,7 +82,7 @@ class StandardSAE(BaseSAE):
         return self.reshape_fn_out(sae_out_pre, self.d_head)
 
 
-class StandardTrainingSAE(BaseTrainingSAE):
+class StandardTrainingSAE(TrainingSAE):
     """
     StandardTrainingSAE is a concrete implementation of BaseTrainingSAE using the "standard" SAE architecture.
     It implements:
@@ -91,10 +93,12 @@ class StandardTrainingSAE(BaseTrainingSAE):
       - calculate_aux_loss: computes a sparsity penalty based on the (optionally scaled) p-norm of feature activations.
     """
 
+    b_enc: nn.Parameter
+
     def initialize_weights(self) -> None:
         # Basic init
         # In Python MRO, this calls StandardSAE.initialize_weights()
-        StandardSAE.initialize_weights(self)
+        StandardSAE.initialize_weights(self)  # type: ignore
 
         # Complex init logic from original TrainingSAE
         if self.cfg.decoder_orthogonal_init:
@@ -118,39 +122,15 @@ class StandardTrainingSAE(BaseTrainingSAE):
         self.W_dec.data /= torch.norm(self.W_dec.data, dim=1, keepdim=True)  # type: ignore
         self.W_dec.data *= norm
 
-    @torch.no_grad()
-    def set_decoder_norm_to_unit_norm(self):
-        self.W_dec.data /= torch.norm(self.W_dec.data, dim=1, keepdim=True)  # type: ignore
-
-    def encode(
-        self, x: Float[torch.Tensor, "... d_in"]
-    ) -> Float[torch.Tensor, "... d_sae"]:
-        # For inference, simply compute feature activations (dropping the pre-activation values)
-        feature_acts, _ = self.encode_with_hidden_pre(x)
-        return feature_acts
-
-    def decode(
-        self, feature_acts: Float[torch.Tensor, "... d_sae"]
-    ) -> Float[torch.Tensor, "... d_in"]:
-        """
-        Decode the feature activations (with the same hooking and normalization as the old SAE).
-        Applies hook_z reshaping if enabled.
-        """
-        scaled_features = self.apply_finetuning_scaling_factor(feature_acts)
-        sae_out_pre = scaled_features @ self.W_dec + self.b_dec
-        sae_out_pre = self.hook_sae_recons(sae_out_pre)
-        sae_out_pre = self.run_time_activation_norm_fn_out(sae_out_pre)
-        return self.reshape_fn_out(sae_out_pre, self.d_head)
-
     def encode_with_hidden_pre(
         self, x: Float[torch.Tensor, "... d_in"]
-    ) -> Tuple[Float[torch.Tensor, "... d_sae"], Float[torch.Tensor, "... d_sae"]]:
+    ) -> tuple[Float[torch.Tensor, "... d_sae"], Float[torch.Tensor, "... d_sae"]]:
         # Process the input (including dtype conversion, hook call, and any activation normalization)
         sae_in = self.process_sae_in(x)
         # Compute the pre-activation (and allow for a hook if desired)
         hidden_pre = self.hook_sae_acts_pre(sae_in @ self.W_enc + self.b_enc)  # type: ignore
         # Add noise during training for robustness (scaled by noise_scale from the configuration)
-        if self.training:
+        if self.training and self.cfg.noise_scale > 0:
             hidden_pre_noised = (
                 hidden_pre + torch.randn_like(hidden_pre) * self.cfg.noise_scale
             )
@@ -162,11 +142,10 @@ class StandardTrainingSAE(BaseTrainingSAE):
 
     def calculate_aux_loss(
         self,
+        step_input: TrainStepInput,
         feature_acts: torch.Tensor,
         hidden_pre: torch.Tensor,
-        dead_neuron_mask: Optional[torch.Tensor],
-        current_l1_coefficient: float,
-        **kwargs: Any,
+        sae_out: torch.Tensor,
     ) -> dict[str, torch.Tensor]:
         # The "standard" auxiliary loss is a sparsity penalty on the feature activations
         weighted_feature_acts = feature_acts
@@ -175,6 +154,14 @@ class StandardTrainingSAE(BaseTrainingSAE):
 
         # Compute the p-norm (set by cfg.lp_norm) over the feature dimension
         sparsity = weighted_feature_acts.norm(p=self.cfg.lp_norm, dim=-1)
-        l1_loss = (current_l1_coefficient * sparsity).mean()
+        l1_loss = (step_input.current_l1_coefficient * sparsity).mean()
 
         return {"l1_loss": l1_loss}
+
+    def log_histograms(self) -> dict[str, NDArray[np.generic]]:
+        """Log histograms of the weights and biases."""
+        b_e_dist = self.b_enc.detach().float().cpu().numpy()
+        return {
+            **super().log_histograms(),
+            "weights/b_e": b_e_dist,
+        }
