+++ tests/training/test_jumprelu_sae.py
@@ -1,20 +1,21 @@
 import pytest
 import torch
 
-from sae_lens.training.training_sae import JumpReLU, TrainingSAE
+from sae_lens.saes.jumprelu_sae import JumpReLU, JumpReLUTrainingSAE, TrainingSAE
+from sae_lens.saes.sae import TrainStepInput
 from tests.helpers import build_sae_cfg
 
 
 def test_jumprelu_sae_encoding():
     cfg = build_sae_cfg(architecture="jumprelu")
-    sae = TrainingSAE.from_dict(cfg.get_training_sae_cfg_dict())
+    sae = JumpReLUTrainingSAE.from_dict(cfg.get_training_sae_cfg_dict())
 
     batch_size = 32
     d_in = sae.cfg.d_in
     d_sae = sae.cfg.d_sae
 
     x = torch.randn(batch_size, d_in)
-    feature_acts, hidden_pre = sae.encode_with_hidden_pre_jumprelu(x)
+    feature_acts, hidden_pre = sae.encode_with_hidden_pre(x)
 
     assert feature_acts.shape == (batch_size, d_sae)
     assert hidden_pre.shape == (batch_size, d_sae)
@@ -39,8 +40,11 @@ def test_jumprelu_sae_training_forward_pass():
 
     x = torch.randn(batch_size, d_in)
     train_step_output = sae.training_forward_pass(
-        sae_in=x,
-        current_l1_coefficient=sae.cfg.l1_coefficient,
+        step_input=TrainStepInput(
+            sae_in=x,
+            current_l1_coefficient=sae.cfg.l1_coefficient,
+            dead_neuron_mask=None,
+        ),
     )
 
     assert train_step_output.sae_out.shape == (batch_size, d_in)
