+++ tests/training/test_sae_trainer.py
@@ -10,6 +10,7 @@ from transformer_lens import HookedTransformer
 from sae_lens import __version__
 from sae_lens.config import LanguageModelSAERunnerConfig
 from sae_lens.sae_training_runner import SAETrainingRunner
+from sae_lens.saes.sae import TrainingSAE
 from sae_lens.training.activations_store import ActivationsStore
 from sae_lens.training.sae_trainer import (
     SAETrainer,
@@ -17,7 +18,6 @@ from sae_lens.training.sae_trainer import (
     _log_feature_sparsity,
     _update_sae_lens_training_version,
 )
-from sae_lens.training.training_sae import TrainingSAE
 from tests.helpers import TINYSTORIES_MODEL, build_sae_cfg, load_model_cached
 
 
@@ -165,9 +165,9 @@ def test_build_train_step_log_dict(trainer: SAETrainer) -> None:
         hidden_pre=torch.tensor([[-1, 0, 0, 1], [1, -1, 0, 1], [1, -1, 1, 1]]).float(),
         loss=torch.tensor(0.5),
         losses={
-            "mse_loss": 0.25,
-            "l1_loss": 0.1,
-            "ghost_grad_loss": 0.15,
+            "mse_loss": torch.tensor(0.25),
+            "l1_loss": torch.tensor(0.1),
+            "ghost_grad_loss": torch.tensor(0.15),
         },
     )
 
@@ -177,23 +177,26 @@ def test_build_train_step_log_dict(trainer: SAETrainer) -> None:
     log_dict = trainer._build_train_step_log_dict(
         output=train_output, n_training_tokens=123
     )
-    assert log_dict == {
-        "losses/mse_loss": 0.25,
-        # l1 loss is scaled by l1_coefficient
-        "losses/l1_loss": train_output.losses["l1_loss"] / trainer.cfg.l1_coefficient,
-        "losses/raw_l1_loss": train_output.losses["l1_loss"],
-        "losses/overall_loss": 0.5,
-        "losses/ghost_grad_loss": 0.15,
-        "metrics/explained_variance": 0.6875,
-        "metrics/explained_variance_legacy": 0.75,
-        "metrics/explained_variance_legacy_std": 0.25,
-        "metrics/l0": 2.0,
-        "sparsity/mean_passes_since_fired": trainer.n_forward_passes_since_fired.mean().item(),
-        "sparsity/dead_features": trainer.dead_neurons.sum().item(),
-        "details/current_learning_rate": 2e-4,
-        "details/current_l1_coefficient": trainer.cfg.l1_coefficient,
-        "details/n_training_tokens": 123,
-    }
+    assert log_dict == pytest.approx(
+        {
+            "losses/mse_loss": 0.25,
+            # l1 loss is scaled by l1_coefficient
+            "losses/l1_loss": train_output.losses["l1_loss"].item()
+            / trainer.cfg.l1_coefficient,
+            "losses/raw_l1_loss": train_output.losses["l1_loss"].item(),
+            "losses/overall_loss": 0.5,
+            "losses/ghost_grad_loss": 0.15,
+            "metrics/explained_variance": 0.6875,
+            "metrics/explained_variance_legacy": 0.75,
+            "metrics/explained_variance_legacy_std": 0.25,
+            "metrics/l0": 2.0,
+            "sparsity/mean_passes_since_fired": trainer.n_forward_passes_since_fired.mean().item(),
+            "sparsity/dead_features": trainer.dead_neurons.sum().item(),
+            "details/current_learning_rate": 2e-4,
+            "details/current_l1_coefficient": trainer.cfg.l1_coefficient,
+            "details/n_training_tokens": 123,
+        }
+    )
 
 
 def test_train_sae_group_on_language_model__runs(
