+++ tests/training/test_upload_saes_to_huggingface.py
@@ -4,7 +4,7 @@ from textwrap import dedent
 import pytest
 from huggingface_hub import HfApi
 
-from sae_lens.sae import SAE
+from sae_lens.saes.sae import SAE
 from sae_lens.training.upload_saes_to_huggingface import (
     _build_sae_path,
     _create_default_readme,
